+++ scripts/continuum_save.sh
@@ -7,7 +7,7 @@ source "$CURRENT_DIR/variables.sh"
 source "$CURRENT_DIR/shared.sh"
 
 supported_tmux_version_ok() {
-	$CURRENT_DIR/check_tmux_version.sh "$SUPPORTED_VERSION"
+	"$CURRENT_DIR/check_tmux_version.sh" "$SUPPORTED_VERSION"
 }
 
 get_interval() {
@@ -29,7 +29,7 @@ enough_time_since_last_run_passed() {
 fetch_and_run_tmux_resurrect_save_script() {
 	local resurrect_save_script_path="$(get_tmux_option "$resurrect_save_path_option" "")"
 	if [ -n "$resurrect_save_script_path" ]; then
-		$resurrect_save_script_path "quiet" >/dev/null 2>&1 &
+		"$resurrect_save_script_path" "quiet" >/dev/null 2>&1 &
 		set_last_save_timestamp
 	fi
 }