+++ scripts/handle_tmux_automatic_start.sh
@@ -17,11 +17,11 @@ is_osx() {
 main() {
 	if is_tmux_automatic_start_enabled; then
 		if is_osx; then
-			$CURRENT_DIR/handle_tmux_automatic_start/osx_enable.sh
+			"$CURRENT_DIR/handle_tmux_automatic_start/osx_enable.sh"
 		fi
 	else
 		if is_osx; then
-			$CURRENT_DIR/handle_tmux_automatic_start/osx_disable.sh
+			"$CURRENT_DIR/handle_tmux_automatic_start/osx_disable.sh"
 		fi
 	fi
 }